add authorization logic in api routes

+++ handlers/recipe.go
@@ -35,6 +35,19 @@ func NewRecipesHandler(ctx context.Context, collection *mongo.Collection, redisC
 
 // CreateRecipeHandler: inserts a new recipe
 func (handler *RecipesHandler) CreateRecipeHandler(c *gin.Context) {
+	// extract the payload from the context that was set by the AuthMiddleware
+	jwtAuthToken, exists := c.Get("auth")
+	if !exists {
+		c.AbortWithStatus(http.StatusUnauthorized)
+		return
+	}
+
+	jwtAuthPayload, ok := jwtAuthToken.(*Claims)
+	if !ok {
+		c.AbortWithStatus(http.StatusUnauthorized)
+		return
+	}
+
 	var recipe models.Recipe
 	err := c.ShouldBindJSON(&recipe)
 	if err != nil {
@@ -43,6 +56,7 @@ func (handler *RecipesHandler) CreateRecipeHandler(c *gin.Context) {
 	}
 
 	recipe.ID = primitive.NewObjectID()
+	recipe.Username = jwtAuthPayload.Username
 	recipe.PublishedAt = time.Now()
 
 	err = handler.recipeService.Create(&recipe)
@@ -93,16 +107,48 @@ func (handler *RecipesHandler) ListRecipesHandler(c *gin.Context) {
 func (handler *RecipesHandler) UpdateRecipeHandler(c *gin.Context) {
 	id := c.Param("id")
 
-	var recipe models.Recipe
-	err := c.ShouldBindJSON(&recipe)
-	if err != nil {
-		c.AbortWithStatusJSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+	// extract the payload from the context that was set by the AuthMiddleware
+	jwtAuthToken, exists := c.Get("auth")
+	if !exists {
+		c.AbortWithStatus(http.StatusUnauthorized)
+		return
+	}
+
+	jwtAuthPayload, ok := jwtAuthToken.(*Claims)
+	if !ok {
+		c.AbortWithStatus(http.StatusUnauthorized)
 		return
 	}
 
 	objectID, err := primitive.ObjectIDFromHex(id)
 	if err != nil {
 		c.AbortWithStatusJSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	var recipe models.Recipe
+	err = c.ShouldBindJSON(&recipe)
+	if err != nil {
+		c.AbortWithStatusJSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	// find a recipe with the requested id
+	recipeRecord, err := handler.recipeService.FindOne(objectID)
+	if err != nil {
+		if err == mongo.ErrNoDocuments {
+			// no recipe record found
+			errMsg := fmt.Sprintf("no recipe found with id: %s", id)
+			c.AbortWithStatusJSON(http.StatusNotFound, gin.H{"error": errMsg})
+			return
+		}
+		c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+
+	if recipeRecord.Username != jwtAuthPayload.Username {
+		errMsg := fmt.Sprintf("you are not the author of the recipe")
+		c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": errMsg})
 		return
 	}
 
@@ -134,6 +180,38 @@ func (handler *RecipesHandler) DeleteRecipeHandler(c *gin.Context) {
 		return
 	}
 
+	// extract the payload from the context that was set by the AuthMiddleware
+	jwtAuthToken, exists := c.Get("auth")
+	if !exists {
+		c.AbortWithStatus(http.StatusUnauthorized)
+		return
+	}
+
+	jwtAuthPayload, ok := jwtAuthToken.(*Claims)
+	if !ok {
+		c.AbortWithStatus(http.StatusUnauthorized)
+		return
+	}
+
+	// find a recipe with the requested id
+	recipe, err := handler.recipeService.FindOne(objectID)
+	if err != nil {
+		if err == mongo.ErrNoDocuments {
+			// no recipe record found
+			errMsg := fmt.Sprintf("no recipe found with id: %s", id)
+			c.AbortWithStatusJSON(http.StatusNotFound, gin.H{"error": errMsg})
+			return
+		}
+		c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+
+	if recipe.Username != jwtAuthPayload.Username {
+		errMsg := fmt.Sprintf("you are not the author of the recipe")
+		c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": errMsg})
+		return
+	}
+
 	recordExists, err := handler.recipeService.Delete(objectID)
 	if err != nil {
 		c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
@@ -160,6 +238,19 @@ func (handler *RecipesHandler) GetOneRecipeHandler(c *gin.Context) {
 		return
 	}
 
+	// extract the payload from the context that was set by the AuthMiddleware
+	jwtAuthToken, exists := c.Get("auth")
+	if !exists {
+		c.AbortWithStatus(http.StatusUnauthorized)
+		return
+	}
+
+	jwtAuthPayload, ok := jwtAuthToken.(*Claims)
+	if !ok {
+		c.AbortWithStatus(http.StatusUnauthorized)
+		return
+	}
+
 	// find a recipe with the requested id
 	recipe, err := handler.recipeService.FindOne(objectID)
 	if err != nil {
@@ -173,6 +264,13 @@ func (handler *RecipesHandler) GetOneRecipeHandler(c *gin.Context) {
 		return
 	}
 
-	c.JSON(http.StatusOK, recipe)
+	// recipe is NOT private or the owner of the recipe themself is fetching the recipe
+	if !recipe.IsPrivate || recipe.Username == jwtAuthPayload.Username {
+		c.JSON(http.StatusOK, recipe)
+		return
+	}
+
+	errMsg := fmt.Sprintf("recipe is private")
+	c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": errMsg})
 	return
 }
